+++ src/main/java/roomescape/controller/GlobalExceptionHandler.java
@@ -8,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.dto.ExceptionResponse;
+import roomescape.controller.dto.response.ExceptionResponse;
 import roomescape.exception.ReservationNotFoundException;
 
 import java.util.stream.Collectors;
